requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.12.0
amplitude-analytics==1.1.4
annotated-types==0.6.0
anyio==4.4.0
asyncpg==0.28.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.39.0
openai-whisper==20231117
peewee==3.17.3
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.20
sympy==1.13.1
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.5
triton==2.3.1
typing_extensions==4.11.0
urllib3==2.2.2
yarl==1.9.4
redis==5.0.8